Began implementing diffuse materials.\nChanged precision to double instead of float

+++ interval.h
@@ -23,7 +23,7 @@ public:
         return min < x && x < max;
     }
 
-    float clamp(float x) const {
+    double clamp(double x) const {
         if (x < min) return min;
         if (x > max) return max;
         return x;
