+++ sphere.h
@@ -6,7 +6,7 @@
 
 class sphere : public hittable {
 public:
-    sphere(const point3& center, float radius) : center(center), radius(std::fmax(0, radius)) {}
+    sphere(const point3& center, double radius) : center(center), radius(std::fmax(0, radius)) {}
 
     bool hit(const ray& r, interval ray_t, hit_record& rec) const override {
         vec3 oc = center - r.origin();
@@ -38,7 +38,7 @@ public:
 
 private:
     point3 center;
-    float radius;
+    double radius;
 };
 
 #endif // SPHERE_H
